fs: rename fs.rs to filesystem.rs
Fixes the following clippy warning:

warning: module has the same name as its containing module
 --> src/fs/mod.rs:8:1
  |
8 | mod fs;
  | ^^^^^^^
  |
  = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#module_inception
  = note: `#[warn(clippy::module_inception)]` on by default
