+++ src/fs/mod.rs
@@ -5,10 +5,10 @@
 // Author: Joerg Roedel <jroedel@suse.de>
 
 mod api;
-mod fs;
+mod filesystem;
 mod init;
 mod ramfs;
 
 pub use api::*;
-pub use fs::*;
+pub use filesystem::*;
 pub use init::populate_ram_fs;
